Release 0.2.0: working USB, refreshed dependencies
USB MIDI now enumerates. It never did before: the device attached, the host
reset it and sent tokens, and the address stayed zero forever. Seven defects,
all in EP0 handling, found by comparing poll/write/read against the Teensy
usb_dev.c interrupt routine:

- enable() never set CONTROL[DPPULLUPNONOTG], so no host saw an attach.
- The USB clock divider was wrong at 120 MHz.
- suspend() asserted USBCTRL[SUSP], which makes the transceiver ignore
  everything but resume signalling, so the host's bus reset was never seen.
- poll() checked SLEEP/RESUME before servicing tokens. A SETUP freezes the
  controller via CTL[TXSUSPENDTOKENBUSY] until firmware clears it, and that
  clear only happens in the token loop, so this livelocked: frozen, so idle,
  so SLEEP, so return early, so still frozen. Tokens are serviced first now,
  and ISTAT[SOFTOK] is cleared so ISTAT can read zero.
- CTL was read-modify-written. TXSUSPENDTOKENBUSY reads as 1 while frozen, so
  modify() fed that bit straight back and re-asserted the freeze, including
  inside reset(). CTL is written whole now, as the reference driver does.
  This is what kept CTL pinned at 0xA1 through every other fix.
- Only the EVEN RX bank was armed, and read() re-armed a separately tracked
  bank. The controller owns the ping-pong sequence, so a software counter
  desynchronises after the first packet and the host's second SETUP lands on
  an unowned bank. Both banks are armed, and read() gives back the bank that
  just completed.
- A SETUP left stale TX descriptors armed, so an IN from an abandoned
  transfer could answer the next IN token. Both EP0 TX descriptors are
  dropped on SETUP. The TX bank pointer is deliberately not reset: the
  controller never consumed those banks, so its own pointer has not moved.

Verified on hardware: enumerates as a USB MIDI device, reaches Configured,
and the host's MIDI stack lists it as a source and a destination.

BREAKING: UsbBusExt::usb_bus() takes &Clocks. The divider was hard-coded to
the 120 MHz values, so a stock 72 MHz Teensy 3.1/3.2 clocked USB at 28.8 MHz
and silently failed to enumerate. It is derived from the configured PLL now
(72, 96 or 120 MHz; anything else panics at init rather than failing quietly).

Dependencies, each a major bump that reaches the public API:

- fugit 0.3 -> 0.6. time::Hertz and PitChannel::start() name fugit types, so
  downstream must move in lockstep or Cargo links two incompatible copies.
  Renames: Rate::raw -> to_raw, Duration::ticks -> as_ticks, millis/secs ->
  from_millis/from_secs.
- defmt 0.3 -> 1. defmt 0.3.100 was already a shim re-exporting defmt 1.0, so
  this is a requirement change with no API change; it drops the shim from
  downstream lockfiles.
- embassy-sync 0.7 -> 0.8. Internal (waitqueue::AtomicWaker, untouched by
  0.8.0) but shared with downstream. Pulls embedded-io-async 0.7, removing a
  duplicate.
- MSRV 1.81 -> 1.82: fugit 0.6 uses const float arithmetic. fugit declares no
  rust-version, so Cargo does not catch this; a 1.81 build fails with a bare
  E0658 inside fugit.

Also: clippy is clean across all three feature configurations. The delay_ns
rounding deliberately keeps `+ 999 / 1000` over div_ceil and allows the lint:
a u64 div_ceil needs quotient and remainder, and with no 64-bit hardware
divide on this core that measured 104 ticks of zero-delay overhead against a
100-tick budget, versus 17 for the original.

Internal notes (CLAUDE.md, STATUS.md, STRATEGY.md) are excluded from the
published package; memory/*.x are linker scripts and still ship.

+++ src/flash.rs
@@ -258,7 +258,7 @@ impl Flash {
         let region = address / PROTECTION_REGION_SIZE;
         // FPROT array in memory: [0x10]=FPROT3 (regions 0-7), [0x11]=FPROT2 (8-15),
         //                        [0x12]=FPROT1 (16-23), [0x13]=FPROT0 (24-31)
-        let fprot_idx = (region / 8) as u32;
+        let fprot_idx = region / 8;
         let fprot_bit = (region % 8) as u8;
         let fprot_val =
             unsafe { ptr::read_volatile((FTFL_BASE + FPROT_OFFSET + fprot_idx) as *const u8) };
